feat: 完成 TUI DDD 架构迁移 (refs #75)
- 收敛 TUI legacy 状态访问与 Effect/update/model 边界
- 补齐 Conversation/Input/Runtime/Session/ViewModel 迁移封口测试
- 清理 runtime/provider/share/tools 兼容接口与 clippy 阻塞
- 更新 #75 TUI Model/View 架构迁移进度

验证：cargo clippy -p cli --all-targets -- -D warnings / cargo test -p cli --all-targets / .agents/hooks/check-unit-tests.sh / cargo check

+++ agent/provider/src/providers/ollama/mod.rs
@@ -140,7 +140,7 @@ impl LlmProvider for OllamaProvider {
 
             if attempt > 0 {
                 let delay =
-                    std::time::Duration::from_millis((1000 * 2u64.pow(attempt as u32)).min(30_000));
+                    std::time::Duration::from_millis((1000 * 2u64.pow(attempt)).min(30_000));
                 log::debug!(
                     "[ollama stream] retry {}/{} after {:?}",
                     attempt,
